+++ src/dapps/RegisterDomain/routes.js
@@ -4,6 +4,7 @@ import InitialENSStateContainer from './containers/InitialENSStateContainer';
 import NameForbiddenENSContainer from './containers/NameForbiddenENSContainer';
 import AlreadyOwnedENSContainer from './containers/AlreadyOwnedENSContainer';
 import FinalizeContainer from './containers/FinalizeContainer';
+import ManageENSContainer from './containers/ManageENSContainer';
 export default {
   path: 'dapps/register-domain',
   component: RegisterDomain,
@@ -49,6 +50,12 @@ export default {
       path: 'finalize',
       name: 'ENS finalize',
       component: FinalizeContainer,
+      props: true
+    },
+    {
+      path: 'manage',
+      name: 'Manage ENS',
+      component: ManageENSContainer,
       props: true
     }
   ]