+++ src/layouts/AccessWalletLayout/components/PasswordModal/PasswordModal.vue
@@ -4,14 +4,16 @@
     :title="$t('accessWallet.password')"
     hide-footer
     class="bootstrap-modal modal-software"
-    centered>
+    centered
+    @shown="focusInput">
     <form class="password-form">
       <div class="input-container">
         <input
+          ref="passwordInput"
           :type="show ? 'text': 'password'"
           v-model="password"
           name="Password"
-          autocomplete="off" >
+          autocomplete="off">
         <img
           v-if="show"
           src="@/assets/images/icons/show-password.svg"
@@ -82,6 +84,9 @@ export default {
     },
     switchViewPassword() {
       this.show = !this.show;
+    },
+    focusInput() {
+      this.$refs.passwordInput.focus();
     }
   }
 };