+++ src/wallets/web3-provider/methods/index.js
@@ -1,0 +1,14 @@
+import ethSendTransaction from './eth_sendTransaction';
+import ethSign from './eth_sign';
+import ethAccounts from './eth_accounts';
+import ethCoinbase from './eth_coinbase';
+import ethSignTransaction from './eth_signTransaction';
+import netVersion from './net_version';
+export {
+  ethSendTransaction,
+  ethSign,
+  ethAccounts,
+  ethCoinbase,
+  ethSignTransaction,
+  netVersion
+};