+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### Feature
+
+- Mnemonic Phrase auto populate [#140](https://github.com/MyEtherWallet/MyEtherWallet/pull/140)
+- Implement Buying Subdomains dapps. [#122](https://github.com/MyEtherWallet/MyEtherWallet/pull/122)
+
+### Devop
+- Refactor web3 provider [#141](https://github.com/MyEtherWallet/MyEtherWallet/pull/141)
+- Refactor Wallet interface [#129](https://github.com/MyEtherWallet/MyEtherWallet/pull/129)
+- Normalise ens-resolver, use resolved address in send container [#126](https://github.com/MyEtherWallet/MyEtherWallet/pull/126)
+
 ### Hotfix
 
 - Some fixes for the ENS Registrar, and validate and normalise ens domains [#120](https://github.com/MyEtherWallet/MyEtherWallet/pull/120)
@@ -6,7 +16,6 @@
 
 ### Devop
 
-- Refactor Wallet interface [#129](https://github.com/MyEtherWallet/MyEtherWallet/pull/129)
 - Use new contract getter [#119](https://github.com/MyEtherWallet/MyEtherWallet/pull/119)
 - Fix access wallet link, pop success when user clicks download [#115](ttps://github.com/MyEtherWallet/MyEtherWallet/pull/115)
 - Add more nodes [#90](ttps://github.com/MyEtherWallet/MyEtherWallet/pull/90)